refactor: stackview contents를 만드는 메서드를 ViewController에서  분리하도록 수정
